Add more to Workspace ajax types (WOR-1818).
