+++ src/workspaces/NewWorkspaceModal/NewWorkspaceModal.tsx
@@ -275,15 +275,15 @@ export const NewWorkspaceModal = withDisplayName(
         Ajax(signal).Groups.list().then(setAllGroups),
         !!cloneWorkspace &&
           Ajax(signal)
-            .Workspaces.workspace(namespace, cloneWorkspace.workspace.name)
+            .Workspaces.workspace(namespace!, cloneWorkspace.workspace.name)
             .details(['workspace.attributes.description'])
             .then((workspace) => {
-              setDescription(workspace.workspace.attributes.description || '');
+              setDescription(workspace.workspace.attributes?.description || '');
             }),
         !!cloneWorkspace &&
           isGoogleWorkspace(cloneWorkspace) &&
           Ajax(signal)
-            .Workspaces.workspace(namespace, cloneWorkspace.workspace.name)
+            .Workspaces.workspace(namespace!, cloneWorkspace.workspace.name)
             .checkBucketLocation(cloneWorkspace.workspace.googleProject, cloneWorkspace.workspace.bucketName)
             .then(({ location }) => {
               // For current phased regionality release, we only allow US or NORTHAMERICA-NORTHEAST1 (Montreal) workspace buckets.
