+++ src/workspaces/utils.ts
@@ -2,16 +2,34 @@ import { cond, safeCurry } from '@terra-ui-packages/core-utils';
 import _ from 'lodash/fp';
 import pluralize from 'pluralize';
 import { AzureBillingProject, BillingProject } from 'src/billing-core/models';
-import { GoogleWorkspaceInfo, WorkspaceInfo, WorkspacePolicy } from 'src/libs/ajax/workspaces/workspace-models';
+import {
+  AzureWorkspace,
+  BaseWorkspace,
+  GoogleWorkspace,
+  GoogleWorkspaceInfo,
+  WorkspaceAccessLevel,
+  workspaceAccessLevels,
+  WorkspaceInfo,
+  WorkspacePolicy,
+  WorkspaceWrapper,
+} from 'src/libs/ajax/workspaces/workspace-models';
 import { azureRegions } from 'src/libs/azure-regions';
 
+export { workspaceAccessLevels } from 'src/libs/ajax/workspaces/workspace-models';
+
 export type {
+  AzureContext,
+  AzureWorkspace,
   AzureWorkspaceInfo,
+  BaseWorkspace,
   BaseWorkspaceInfo,
+  GoogleWorkspace,
   GoogleWorkspaceInfo,
+  WorkspaceAccessLevel,
   WorkspaceInfo,
   WorkspacePolicy,
   WorkspaceState,
+  WorkspaceWrapper,
 } from 'src/libs/ajax/workspaces/workspace-models';
 
 export type CloudProvider = 'AZURE' | 'GCP';
@@ -33,12 +51,6 @@ export const isGoogleWorkspaceInfo = (workspace: WorkspaceInfo | undefined): wor
   return workspace ? workspace.cloudPlatform === 'Gcp' : false;
 };
 
-export const workspaceAccessLevels = ['NO ACCESS', 'READER', 'WRITER', 'OWNER', 'PROJECT_OWNER'] as const;
-
-export type WorkspaceAccessLevels = typeof workspaceAccessLevels;
-
-export type WorkspaceAccessLevel = WorkspaceAccessLevels[number];
-
 export const hasAccessLevel = (required: WorkspaceAccessLevel, current: WorkspaceAccessLevel): boolean => {
   return workspaceAccessLevels.indexOf(current) >= workspaceAccessLevels.indexOf(required);
 };
@@ -46,29 +58,6 @@ export const hasAccessLevel = (required: WorkspaceAccessLevel, current: Workspac
 export const canWrite = (accessLevel: WorkspaceAccessLevel): boolean => hasAccessLevel('WRITER', accessLevel);
 export const canRead = (accessLevel: WorkspaceAccessLevel): boolean => hasAccessLevel('READER', accessLevel);
 export const isOwner = (accessLevel: WorkspaceAccessLevel): boolean => hasAccessLevel('OWNER', accessLevel);
-
-export interface WorkspaceSubmissionStats {
-  lastSuccessDate?: string;
-  lastFailureDate?: string;
-  runningSubmissionsCount: number;
-}
-
-export interface BaseWorkspace {
-  owners?: string[];
-  accessLevel: WorkspaceAccessLevel;
-  canShare: boolean;
-  canCompute: boolean;
-  workspace: WorkspaceInfo;
-  policies: WorkspacePolicy[];
-  public?: boolean;
-  workspaceSubmissionStats?: WorkspaceSubmissionStats;
-}
-
-export interface AzureContext {
-  managedResourceGroupId: string;
-  subscriptionId: string;
-  tenantId: string;
-}
 
 export interface PolicyDescription {
   shortDescription: string;
@@ -106,16 +95,6 @@ export const getPolicyDescriptions = (
   }
   return policyDescriptions;
 };
-
-export interface AzureWorkspace extends BaseWorkspace {
-  azureContext: AzureContext;
-}
-
-export interface GoogleWorkspace extends BaseWorkspace {
-  workspace: GoogleWorkspaceInfo;
-}
-
-export type WorkspaceWrapper = GoogleWorkspace | AzureWorkspace;
 
 export const isAzureWorkspace = (workspace: BaseWorkspace): workspace is AzureWorkspace => {
   return workspace.workspace.cloudPlatform === 'Azure';
